now blog can be deleted only by the user who added it (exercise 4.21)

+++ part4/controllers/blogs.js
@@ -34,11 +34,19 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  const blog = await Blog.findByIdAndDelete(request.params.id)
-
+  const blog = await Blog.findById(request.params.id)
   if (!blog) {
     return response.status(404).json({ error: 'Blog not found' })
   }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id)
+    return response.status(401).json({ error: 'Invalid token' })
+
+  if (blog.user.toString() !== decodedToken.id)
+    return response.status(403).json({ error: 'You do not have permission to delete this blog' })
+
+  await Blog.findByIdAndDelete(request.params.id)
 
   response.status(204).end()
 })
